fix: deployment process for Linux is now in Docker

+++ etc/scripts/build-pyarmored-dist.sh
@@ -8,6 +8,7 @@ OUTPUT_DIR="./dist/pyarmor"
 TMP_DIR="./tmp"
 OBFUSCATED_PACKAGES="aiocflib flockwave skybrush"
 TARBALL_NAME="skybrush-server"
+VENV_DIR="${VENV_DIR:-.venv}"
 
 ###############################################################################
 
@@ -19,6 +20,11 @@ STANDALONE=0
 
 if [ "x$1" = "x--standalone" ]; then
   STANDALONE=1
+  shift
+fi
+
+if [ "x$1" != x ]; then
+  VENV_DIR="$1"
 fi
 
 cd "${REPO_ROOT}"
@@ -27,22 +33,34 @@ cd "${REPO_ROOT}"
 PROJECT_NAME=`cat pyproject.toml|grep ^name|head -1|cut -d '"' -f 2`
 VERSION=`cat pyproject.toml|grep ^version|head -1|cut -d '"' -f 2`
 
-# Remove all requirements.txt files, we don't use them, only poetry
-rm -f requirements*.txt
+POETRY=`which poetry 2>/dev/null || true`
 
-# Build the Skybrush source tarball first
-rm -rf dist/"${PROJECT_NAME}"*.tar.gz
-poetry build -f sdist
+if [ "x${POETRY}" != x ]; then
+  # Remove all requirements.txt files, we don't use them, only poetry
+  rm -f requirements*.txt
 
-# Generate requirements.txt from poetry
-poetry export -f requirements.txt -o requirements.txt --without-hashes --with-credentials
-ls dist/${PROJECT_NAME}*.tar.gz >>requirements.txt
-trap "rm -f requirements.txt" EXIT
+  # Generate requirements.txt from poetry. We use requirements-main.txt for sake of
+  # compatibility with building this in a pyinstaller Docker container where we
+  # cannot use requirements.txt
+  "${POETRY}" export -f requirements.txt -o requirements-main.txt --without-hashes --with-credentials
+  trap "rm -f requirements-main.txt" EXIT
+
+  # Build the Skybrush wheel and append it to the requirements
+  rm -rf dist/"${PROJECT_NAME}"*.whl
+  "${POETRY}" build -f wheel
+  ls dist/*.whl >>requirements-main.txt
+else
+  echo "Poetry not installed; we assume that the requirements are already prepared in requirements-main.txt"
+fi
 
 # Create virtual environment if it doesn't exist yet
-if [ ! -d .venv ]; then
-  python3 -m venv .venv
+if [ ! -d "${VENV_DIR}" ]; then
+  python3 -m venv "${VENV_DIR}"
 fi
+PIP="${VENV_DIR}/bin/pip"
+PYARMOR="${VENV_DIR}/bin/pyarmor"
+PYINSTALLER="${VENV_DIR}/bin/pyinstaller"
+PYTHON="${VENV_DIR}/bin/python"
 
 # Create build folder
 rm -rf "${BUILD_DIR}"
@@ -50,8 +68,8 @@ mkdir -p "${BUILD_DIR}/bin"
 mkdir -p "${BUILD_DIR}/lib"
 
 # Install dependencies
-.venv/bin/pip install -U pip wheel "pyarmor>=6.6.0" pyinstaller
-.venv/bin/pip install -r requirements.txt -t "${BUILD_DIR}/lib"
+"${PIP}" install -U pip wheel "pyarmor>=6.6.0" pyinstaller
+"${PIP}" install -r requirements-main.txt -t "${BUILD_DIR}/lib"
 
 # lxml is huge and we don't need it; pymavlink brings it in mistakenly as a
 # dependency
@@ -75,15 +93,17 @@ if [ $STANDALONE = 1 ]; then
   # Invoke obfuscation script on virtualenv. Note that we need --advanced 2 here
   # because the PyInstaller repacking trick does not work with the "normal"
   # mode.
-  PYARMOR_ARGS="--advanced 2" etc/scripts/apply-pyarmor-on-venv.sh .venv/bin/pyarmor "${BUILD_DIR}/lib" "${BUILD_DIR}/obf" --keep
+  PYARMOR_ARGS="--advanced 2" etc/scripts/apply-pyarmor-on-venv.sh "${PYARMOR}" "${BUILD_DIR}/lib" "${BUILD_DIR}/obf" --keep
 
   # Call PyInstaller to produce an unobfuscated distribution first
   # TODO(ntamas): BUILD_DIR is hardcoded into pyinstaller.spec
-  .venv/bin/pyinstaller --clean -y --dist "${BUILD_DIR}/dist" etc/deployment/pyinstaller/pyinstaller.spec
+  # Apparently we need to pass PYTHONPATH earlier because pkg_resources does not
+  # find stuff in it if we let PyInstaller extend the PATH later
+  PYTHONPATH="${BUILD_DIR}/lib" "${PYINSTALLER}" --clean -y --dist "${BUILD_DIR}/dist" etc/deployment/pyinstaller/pyinstaller.spec
 
   # Replace the unobfuscated libraries with the obfuscated ones in the PyInstaller
   # distribution
-  .venv/bin/python etc/deployment/pyarmor/repack.py -p "${BUILD_DIR}/obf" "${BUILD_DIR}/dist/skybrushd"
+  "${PYTHON}" etc/deployment/pyarmor/repack.py -p "${BUILD_DIR}/obf" "${BUILD_DIR}/dist/skybrushd"
   mv skybrushd_obf ${BUILD_DIR}/bin/skybrushd
 
   # Clean up after ourselves
@@ -91,14 +111,31 @@ if [ $STANDALONE = 1 ]; then
   rm -rf "${BUILD_DIR}/dist"
   rm -rf "${BUILD_DIR}/lib"
   rm -rf "${BUILD_DIR}/obf"
+
+  # Okay, so in the build dir we now have doc/* with the license files and bin/
+  # with the obfuscated bundle. Let's rearrange stuff a bit so we have a place for
+  # the config file and the license file
+  rm -rf "${BUILD_DIR}/staging"
+  mkdir -p "${BUILD_DIR}/staging/bin"
+  mv "${BUILD_DIR}/bin" "${BUILD_DIR}/staging/lib"
+  mv "${BUILD_DIR}/doc" "${BUILD_DIR}/staging/doc"
+
+  # TODO(ntamas): create separate bundles for indoor and outdoor
+  cp etc/deployment/configs/skybrush-indoor.jsonc "${BUILD_DIR}/staging/skybrush.jsonc"
+  cp etc/deployment/linux/skybrushd "${BUILD_DIR}/staging/bin/skybrushd"
+  chmod a+x "${BUILD_DIR}/staging/bin/skybrushd"
 else
   # Separate obfuscated Python modules; typically for a Raspberry Pi
 
   # Invoke obfuscation script on virtualenv
-  etc/scripts/apply-pyarmor-on-venv.sh .venv/bin/pyarmor "${BUILD_DIR}/lib" "${BUILD_DIR}/obf"
+  etc/scripts/apply-pyarmor-on-venv.sh "${PYARMOR}" "${BUILD_DIR}/lib" "${BUILD_DIR}/obf"
 
   # Create a launcher script
   cp etc/deployment/pyarmor/skybrushd "${BUILD_DIR}/bin"
+
+  # Move everything to the staging dir
+  mkdir -p "${BUILD_DIR}/staging"
+  mv "${BUILD_DIR}/bin" "${BUILD_DIR}/lib" "${BUILD_DIR}/staging"
 fi
 
 # Create a tarball
@@ -106,9 +143,10 @@ TARBALL_STEM="${TARBALL_NAME}-${VERSION}"
 rm -rf "${TMP_DIR}/${TARBALL_STEM}"
 mkdir -p "${TMP_DIR}/${TARBALL_STEM}"
 mkdir -p "${OUTPUT_DIR}"
-mv "${BUILD_DIR}"/* "${TMP_DIR}/${TARBALL_STEM}"
-tar -C "${TMP_DIR}" --exclude "__pycache__" -czf "${OUTPUT_DIR}/${TARBALL_STEM}.tar.gz" "${TARBALL_STEM}/"
+mv ${BUILD_DIR}/staging/* "${TMP_DIR}/${TARBALL_STEM}"
+tar -C "${TMP_DIR}" --owner=0 --group=0 --exclude "__pycache__" -czf "${OUTPUT_DIR}/${TARBALL_STEM}.tar.gz" "${TARBALL_STEM}/"
 rm -rf "${TMP_DIR}/${TARBALL_STEM}"
+rm -rf "${BUILD_DIR}/staging"
 
 echo ""
 echo "------------------------------------------------------------------------"
